GESIS Search

Query the GESIS search engine to retrieve metadata on research data, variables, tools, publications, libraries, and websites.

Usage

gesis_search(
  query = NULL,
  type = NULL,
  default_operator = "AND",
  fields = field_weights(),
  sort = "relevance",
  publication_year = NULL,
  countries = NULL,
  topics = NULL,
  authors = NULL,
  sources = NULL,
  study_title = NULL,
  study_group = NULL,
  collection_year = NULL,
  thematic_collections = NULL,
  method_data_collection = NULL,
  method_sampling = NULL,
  method_time = NULL,
  method_unit = NULL,
  data_type = NULL,
  interview_language = NULL,
  pages = 1,
  tidy = FALSE
)

# S3 method for class 'gesis_hits'
as.data.frame(x, ...)

Arguments

query

The search term to query. Terms can be seperated by operators OR / AND. If no operator is specified, defaults to the operator specified in default_operator. If NULL, searches for all records. Queryless searches can still be filtered using the parameters below.

type

Type of search results to retrieve. Can be one of "research_data", "variables", "instruments_tools", "publication", "gesis_bib", or "gesis_web" or "gesis_person". If NULL, includes all result types. Defaults to NULL.

default_operator

Character specifying the default operator to use as seperator of search terms. Can be "AND" or "OR". Defaults to "AND". Ignored if query is NULL.

fields

Character vector of field weights as returned by field_weights. Used to specify how much weight is given to query matches in certain metadata fields. The higher the weight, the more a field influences the match probability. Defaults to the default values specified by field_weights. Ignored if query is NULL.

sort

Attribute to sort by. Must be one of "relevance", "newest_first", "oldest_first", "title_ascending", "title_descending", "variable_ascending", or "variable_descending". Defaults to "relevance".

publication_year

Numeric or character vector of length 2 specifying the time range in which a record was published. The first value is the start year and the second value is the end year. If NULL, returns records from all years. Defaults to NULL.

countries

Character vector specifying the country in which the record was published. If NULL, returns records from all countries. Defaults to NULL.

topics

Character vector specifying the topics to filter by. Example values include "Political behavior and attitudes" or "Migration". See https://search.gesis.org/ for a full list of possible topics.

authors

Character vector specifying the authors to filter by.

sources

Character vector specifying the source of the record.

study_title

Character vector specifying the study title. Only sensible for research data or variables.

study_group

Character vector specifying the study group. Only sensible for research data or variables.

collection_year

Numeric or character vector of length 2 specifying the time range in which a dataset was collected. The first value is the start year and the second value is the end year. If NULL, returns records from all years. Defaults to NULL. Only sensible for research data or variables.

thematic_collections

Character vector specifying the thematic collection a dataset or variable belongs to. Example values include "COVID-19" or "Environment and climate". See https://search.gesis.org/ for a full list of possible topics.

method_data_collection, method_sampling, method_time, method_unit

Arguments to filter by the methodology used in a study. It is possible to filter by the mode of data collection (method_data_collection), sampling procedure (method_sampling), temporal research design (method_unit) and analysis unit (method_unit). Only sensible for research data and variables.

data_type

Character vector specifying the data type of a dataset or variable. Example values include "numeric", "geospatial", and "text". Only sensible for research data and variables.

interview_language

Character vector specifying the interview language of a variable. Usually only "german" and "english" are available. Only sensible for variables.

pages

Numeric vector specifying the page numbers of the search results. By default, only the first page is returned. If NULL, iterates through all pages.

tidy

If TRUE, returns the records in a rectangular shape that can be wrangled more easily. In this form, some more complex metadata fields that cannot be converted to a rectangular shapre are discarded. If FALSE, returns the records as a list-like gesis_hits object. This object is more comprehensive, but harder to work with. Defaults to FALSE.

x

Object of class gesis_hits

...

Arguments passed on to the respective S3 methods.

Value

If tidy = FALSE, returns an object of class gesis_hits consisting of multiple objects of class gesis_hit. Each gesis_hit contains attributes index, id, and score specifying the index, the record ID, and the match score.

If tidy = TRUE, returns a dataframe or tibble where each row is a record and each column is a metadata field.

In any case, the query used to retrieve the data is stored in the attribute query.

Examples

# Make a queryless search across the entire database and return
# the first 10 results
gesis_search()
#> A list of <gesis_hits> with 10 hits
#> 
#> <gesis_hit>
#> Type: publication
#> ID: gesis-ssoar-22993
#> Title: Globalization and the Housing Asset Rich: Geographies, Demographies and
#> Policy Convoys
#> Date: 2008
#> Person:
#> • Forrest, Ray
#> 
#> <gesis_hit>
#> Type: publication
#> ID: gesis-ssoar-22994
#> Title: From Structural Adjustment to Social Adjustment
#> Date: 2008
#> Person:
#> • Bradshaw, Sarah
#> 
#> <gesis_hit>
#> Type: publication
#> ID: gesis-ssoar-22998
#> Title: Globalization, Conflict and Mental Health
#> Date: 2008
#> Person:
#> • Piachaud, Jack
#> # ℹ 7 more hits
#> # ℹ Use `print(n = ...)` to see more hits

# Queries can be narrowed down by providing a query string
hit <- gesis_search("climate change")

# gesis_search() searches for all types of records by default. If you only
# want to search for, say, datasets, you can set the `type` argument.
gesis_search(type = "research_data")
#> A list of <gesis_hits> with 10 hits
#> 
#> <gesis_hit>
#> Type: research_data
#> ID: SDN-10.7802-2333
#> Title: Die Themen "Umwelt" und "Corona" in der medialen Berichterstattung
#> während der ersten Coronawelle in Deutschland 2020
#> Date: 2021
#> Person:
#> • Krug, Maximilian
#> 
#> <gesis_hit>
#> Type: research_data
#> ID: SDN-10.7802-2433
#> Title: HMC Community Survey 2021
#> Date: 2022
#> Persons:
#> • Arndt, Witold
#> • Gerlich, Silke Christine
#> • Hofmann, Volker
#> • Kubin, Markus
#> • Kulla, Lucas
#> • ... and 9 more
#> 
#> <gesis_hit>
#> Type: research_data
#> ID: SDN-10.7802-2508
#> Title: SiE survey dataset on solidarity in Europe (2021)
#> Date: 2021
#> Persons:
#> • Hemerijck, Anton
#> • Genschel, Philipp
#> • Cicchi, Lorenzo
#> • Nasr, Mohamed
#> # ℹ 7 more hits
#> # ℹ Use `print(n = ...)` to see more hits

# By default, only 10 results (1 page) are returned. You can specify
# which pages should queried or if all should be returned
gesis_search("climate change", pages = 2:3) # pages 2 and 3
#> A list of <gesis_hits> with 20 hits
#> 
#> <gesis_hit>
#> Type: gesis_bib
#> ID: gesis-bib-126747
#> Title: Climate change and cultural diversity
#> Person:
#> • Wisner, Ben
#> 
#> <gesis_hit>
#> Type: publication
#> ID: gesis-ssoar-79316
#> Title: Germany's Federal Climate Change Act
#> Date: 2021
#> Persons:
#> • Flachsland, Christian
#> • Levi, Sebastian
#> 
#> <gesis_hit>
#> Type: publication
#> ID: gesis-ssoar-45173
#> Title: Can climate change negotiations succeed?
#> Date: 2013
#> Persons:
#> • Hovi, Jon
#> • Skodvin, Tora
#> • Aakre, Stine
#> # ℹ 17 more hits
#> # ℹ Use `print(n = ...)` to see more hits
if (FALSE) gesis_search("climate change", pages = NULL) # all pages

# By default, results are wrapped in a complex list structure. This is
# good for representing detailed information, but not for analyzing data.
# Results can be tidied up to a dataframe using either `as.data.frame()`
# or by setting the `tidy` argument to TRUE
as.data.frame(hit)
#>                                 id
#> 1             csa-pais-2011-148719
#> 2                gesis-ssoar-75571
#> 3                 gesis-bib-142488
#> 4                   fes-bib-469114
#> 5                 gesis-bib-145021
#> 6                gesis-ssoar-76642
#> 7  bibsonomy-goughietal2008climate
#> 8                 gesis-bib-134347
#> 9                 gesis-bib-153820
#> 10             gesis-bib-153820-LP
#>                                                               title        type
#> 1                                      Migration and Climate Change publication
#> 2  Japan's Climate Change Discourse: Toward Climate Securitisation? publication
#> 3                                      The climate change imaginary   gesis_bib
#> 4                                   Human rights and climate change publication
#> 5                             Future generations and climate change   gesis_bib
#> 6                         Attitudes towards climate change migrants publication
#> 7                                  Climate Change and Social Policy publication
#> 8                            Introduction to climate change justice   gesis_bib
#> 9                            Climate Change : Fieldwork: April 2019   gesis_bib
#> 10                           Climate Change : Fieldwork: April 2019 publication
#>    date date_recency
#> 1  2011         2011
#> 2  2021         2021
#> 3  <NA>         <NA>
#> 4  2009         2009
#> 5  <NA>         <NA>
#> 6  2020         2020
#> 7  2008         2008
#> 8  <NA>         <NA>
#> 9  2019         2019
#> 10 2019         2019
#>                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                               abstract
#> 1                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                         Provides an overview of the relationship between climate change and migration. The study brings together both case studies and syntheses from different parts of the world and discusses empirical evidence, methodological challenges, conceptual gaps, policy responses, and normative issues. The book constitutes an introduction to one of the most discussed but least understood consequences of climate change and brings together contributors from a wide range of disciplines, including anthropology, climatology, demography, geography, law, political science and sociology. Tables, Figures.
#> 2                         This article situates Japan in the international climate security debate by analysing competing climate change discourses. In 2020, for the first time, the Japanese Ministry of the Environment included the term “climate crisis” (kikō kiki) in its annual white paper, and the Japanese parliament adopted a “climate emergency declaration” (kikō hijō jitai sengen). Does this mean that Japan’s climate discourse is turning toward the securitisation of climate change? Drawing on securitisation theory, this article investigates whether we are seeing the emergence of a climate change securitisation discourse that treats climate change as a security issue rather than a conventional political issue. The analysis focuses on different stakeholders in Japan’s climate policy: the Japanese Ministry of the Environment, the Ministry of Economy, Trade, and Industry, the Ministry of Foreign Affairs, the parliament, the Cabinet, and sub- and non-state actors. Through a discourse analysis of ministry white papers and publications by other stakeholders, the article identifies a burgeoning securitisation discourse that challenges, albeit moderately, the status quo of incrementalism and inaction in Japan’s climate policy. This article further highlights Japan’s position in the rapidly evolving global debate on the urgency of climate action and provides explanations for apparent changes and continuities in Japan’s climate change discourse.
#> 3                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 <NA>
#> 4                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 <NA>
#> 5                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 <NA>
#> 6  While climate change has become a salient political and social issue in Western societies, we know relatively little about how these societies will react to one of the potentially important consequences of climate change: increasing migration flows. By means of a representative online survey in Germany, this paper therefore investigates for the very first time to what extent citizens in industrialized societies are willing to accept climate change migrants, especially in comparison with other groups of migrants and refugees, and the circumstances and principles under which they would accept them. The findings show that climate change migrants receive high support levels comparable with those enjoyed by political refugees (migrants who need special protection) and that contrast with attitudes towards economic migrants (who are often not seen as in need of special protection). We also see that people are more likely to accept justifications for taking climate change migrants when they realize that the expected number of migrants is relatively low. While arguments about morality, corrective justice, and a country’s capacity lead to similar acceptance rates in general, the latter argument plays a more important role for highly educated people and non-environmentalists than the former two. The findings of this study allow us to better prepare for potential conflicts that might emerge with increasing migration flows caused by climate change.
#> 7                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                     
#> 8                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 <NA>
#> 9                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                 <NA>
#> 10                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                <NA>
#>                                                  portal_url
#> 1  http://sowiport.gesis.org/search/id/csa-pais-2011-148719
#> 2        https://www.ssoar.info/ssoar/handle/document/75571
#> 3                                                      <NA>
#> 4        http://sowiport.gesis.org/search/id/fes-bib-469114
#> 5                                                      <NA>
#> 6        https://www.ssoar.info/ssoar/handle/document/76642
#> 7                                                      <NA>
#> 8                                                      <NA>
#> 9                                                      <NA>
#> 10                                                     <NA>
#>                                                                                                                                                                                                                                    topic
#> 1                                                                                                                                                           Global warming, Migration, Law, Environment, Political science, Anthropology
#> 2                                                                                                                                                                                       Japan, Klimawandel, Krise, Diskurs, Klimapolitik
#> 3                                                                                                                                                                                                                   Zeitschriftenartikel
#> 4                                                                                                                                                                                                            Klimaschaden, Menschenrecht
#> 5                                                                                                                                                                                                                   Zeitschriftenartikel
#> 6                                                                                                                                  Klimawandel, Migration, Flucht, Bundesrepublik Deutschland, öffentliche Meinung, Akzeptanz, Vorurteil
#> 7                                                                                                                                                                                                                                     NA
#> 8                                                                                                                                                                                                                   Zeitschriftenartikel
#> 9  Energieverbrauch, erneuerbare Energie, EU, Klimaschutz, Klimawandel, Konsumverhalten, öffentliche Meinung, Umweltbewusstsein, Wahrnehmung, pdf, GESIS-Studie, FDZ International, Umfrageergebnisse, Abschlussbericht, Primärliteratur
#> 10 Energieverbrauch, erneuerbare Energie, EU, Klimaschutz, Klimawandel, Konsumverhalten, öffentliche Meinung, Umweltbewusstsein, Wahrnehmung, pdf, GESIS-Studie, FDZ International, Umfrageergebnisse, Abschlussbericht, Primärliteratur
#>                                                                   person
#> 1  Piguet, Etienne [Ed], Pecoud, Antoine [Ed], de Guchteneire, Paul [Ed]
#> 2                                   Koppenborg, Florentine, Hanssen, Ulv
#> 3                                                       Luke, Timothy W.
#> 4                                                                     NA
#> 5                                                        Nelson, Gillian
#> 6                                                         Helbling, Marc
#> 7                                                        Gough, I. et al
#> 8                                                           Brooks, Thom
#> 9                                                                     NA
#> 10                                                                    NA
#>                   person_sort
#> 1  PiguetPecoudde Guchteneire
#> 2           KoppenborgHanssen
#> 3                        Luke
#> 4                        <NA>
#> 5                      Nelson
#> 6                    Helbling
#> 7                       Gough
#> 8                      Brooks
#> 9                        <NA>
#> 10                       <NA>
#>                                                                                   source
#> 1                                                       Cambridge University Press, 2011
#> 2                                         In: Politics and Governance, 9, 2021, 4, 53-64
#> 3                                In: Current Sociology, vol. 63(2015) no. 2 ; p. 280-296
#> 4                                                                    Geneva, 2009, 34 S.
#> 5  In: International Social Science Journal, 2013 no. 211/212 ; p. 89-98. ISSN 0020-8701
#> 6                                              In: Climatic Change, 160, 2020, 1, 89-102
#> 7                                In Journal of European Social Policy, 18, 325-344, 2008
#> 8                        In: Political Science & Politics, vol. 46(2013) no. 1 ; p. 9-12
#> 9                      Brussels: European Commission, 2019.- 126 S. ; Tab., graf. Darst.
#> 10                     Brussels: European Commission, 2019.- 126 S. ; Tab., graf. Darst.
#>            subtype publisher                                      database
#> 1             book  ProQuest     PAIS - Public Affairs Information Service
#> 2  journal_article      <NA> SSOAR - Social Science Open Access Repository
#> 3  journal_article     GESIS                            GESIS - Bibliothek
#> 4             book       FES        FES - Friedrich-Ebert-Stiftung Katalog
#> 5  journal_article     GESIS                            GESIS - Bibliothek
#> 6  journal_article      <NA> SSOAR - Social Science Open Access Repository
#> 7             <NA>      <NA>                                          <NA>
#> 8  journal_article     GESIS                            GESIS - Bibliothek
#> 9             book     GESIS                            GESIS - Bibliothek
#> 10            book     GESIS                            GESIS - Bibliothek
#>           document_type
#> 1                  Buch
#> 2  Zeitschriftenartikel
#> 3  Zeitschriftenaufsatz
#> 4                  Buch
#> 5  Zeitschriftenaufsatz
#> 6  Zeitschriftenartikel
#> 7               article
#> 8  Zeitschriftenaufsatz
#> 9                  Buch
#> 10                 Buch
#>                                                               coreAuthor
#> 1  Piguet, Etienne [Ed], Pecoud, Antoine [Ed], de Guchteneire, Paul [Ed]
#> 2                                   Koppenborg, Florentine, Hanssen, Ulv
#> 3                                                       Luke, Timothy W.
#> 4                                                                     NA
#> 5                                                        Nelson, Gillian
#> 6                                                         Helbling, Marc
#> 7                                                                     NA
#> 8                                                           Brooks, Thom
#> 9                                                                     NA
#> 10                                                                    NA
#>                 corePublisher      coreIsbn       coredoctypelitadd
#> 1  Cambridge University Press 9781107014855            Keine Angabe
#> 2                        <NA>          <NA>                    <NA>
#> 3                        <NA>          <NA>             Monographie
#> 4                        <NA>          <NA>                    <NA>
#> 5                        <NA>          <NA>             Monographie
#> 6                        <NA>          <NA>                    <NA>
#> 7                                      <NA>                    <NA>
#> 8                        <NA>          <NA>             Monographie
#> 9         European Commission 9789276093367 Elektronische Ressource
#> 10        European Commission 9789276093367 Elektronische Ressource
#>     coreLanguage         data_source link_count gesis_own fulltext index_source
#> 1  Englisch (EN) GESIS-Literaturpool          0         1        0      OUTCITE
#> 2             en         GESIS-SSOAR          0         1        1  GESIS-SSOAR
#> 3  Englisch (EN)           GESIS-BIB       <NA>         1        0    GESIS-BIB
#> 4   Keine Angabe GESIS-Literaturpool          0         1        1      OUTCITE
#> 5  Englisch (EN)           GESIS-BIB       <NA>         1        0    GESIS-BIB
#> 6             en         GESIS-SSOAR          0         1        1  GESIS-SSOAR
#> 7           <NA>   ISSP-Bibliography       <NA>         1     <NA>    Bibsonomy
#> 8  Englisch (EN)           GESIS-BIB       <NA>         1        0    GESIS-BIB
#> 9  Englisch (EN)           GESIS-BIB          1         1        1    GESIS-BIB
#> 10 Englisch (EN)           GESIS-BIB          1         1        1  LINK-DB-NEW
#>    literature_collection coreSjahr                     coreJournalTitle
#> 1    GESIS-Literaturpool      <NA>                                 <NA>
#> 2                   <NA>      2021              Politics and Governance
#> 3                   <NA>      <NA>                    Current Sociology
#> 4    GESIS-Literaturpool      <NA>                                 <NA>
#> 5                   <NA>      <NA> International Social Science Journal
#> 6                   <NA>      2020                      Climatic Change
#> 7    GESIS-Literaturpool      <NA>    Journal of European Social Policy
#> 8                   <NA>      <NA>         Political Science & Politics
#> 9                   <NA>      <NA>                                 <NA>
#> 10   GESIS-Literaturpool      <NA>                                 <NA>
#>    coreZsband coreZsnummer                                        doi  coreIssn
#> 1        <NA>         <NA>                                       <NA>      <NA>
#> 2           9            4     https://doi.org/10.17645/pag.v9i4.4419 2183-2463
#> 3        <NA>         <NA>                                       <NA>      <NA>
#> 4        <NA>         <NA>                                       <NA>      <NA>
#> 5        <NA>         <NA>                                       <NA>      <NA>
#> 6         160            1 https://doi.org/10.1007/s10584-020-02697-3 1573-1480
#> 7          18                                                    <NA>      <NA>
#> 8        <NA>         <NA>                                       <NA>      <NA>
#> 9        <NA>         <NA>                                       <NA>      <NA>
#> 10       <NA>         <NA>                                       <NA>      <NA>
#>    metadata_quality             library_location
#> 1              <NA>                         <NA>
#> 2                10                         <NA>
#> 3              <NA>    Köln Zeitschriftenaufsatz
#> 4              <NA>                         <NA>
#> 5              <NA>    Köln Zeitschriftenaufsatz
#> 6                10                         <NA>
#> 7              <NA>                         <NA>
#> 8              <NA>    Köln Zeitschriftenaufsatz
#> 9              <NA> Köln Elektronische Ressource
#> 10             <NA> Köln Elektronische Ressource
#>                                institutions
#> 1                                      <NA>
#> 2                                      <NA>
#> 3                                      <NA>
#> 4  Friedrich-Ebert-Stiftung / Büro <Genève>
#> 5                                      <NA>
#> 6                                       WZB
#> 7                                      <NA>
#> 8                                      <NA>
#> 9                                      <NA>
#> 10                                     <NA>
#>                              coreCorpEditor coreEditor coreSatit corePagerange
#> 1                                      <NA>       <NA>      <NA>          <NA>
#> 2                                      <NA>       <NA>      <NA>          <NA>
#> 3                                      <NA>       <NA>      <NA>          <NA>
#> 4  Friedrich-Ebert-Stiftung / Büro <Genève>       <NA>      <NA>          <NA>
#> 5                                      <NA>       <NA>      <NA>          <NA>
#> 6                                      <NA>       <NA>      <NA>          <NA>
#> 7                                      <NA>                            325-344
#> 8                                      <NA>       <NA>      <NA>          <NA>
#> 9                                      <NA>       <NA>      <NA>          <NA>
#> 10                                     <NA>       <NA>      <NA>          <NA>
#>    institution            bibtex_key remarks address recordurl
#> 1         <NA>                  <NA>    <NA>    <NA>      <NA>
#> 2         <NA>                  <NA>    <NA>    <NA>      <NA>
#> 3         <NA>                  <NA>    <NA>    <NA>      <NA>
#> 4         <NA>                  <NA>    <NA>    <NA>      <NA>
#> 5         <NA>                  <NA>    <NA>    <NA>      <NA>
#> 6         <NA>                  <NA>    <NA>    <NA>      <NA>
#> 7              goughietal2008climate  (ISSP)                  
#> 8         <NA>                  <NA>    <NA>    <NA>      <NA>
#> 9         <NA>                  <NA>    <NA>    <NA>      <NA>
#> 10        <NA>                  <NA>    <NA>    <NA>      <NA>
#>                                                                                                                             tags
#> 1                                                                                                                             NA
#> 2                                                                                                                             NA
#> 3                                                                                                                             NA
#> 4                                                                                                                             NA
#> 5                                                                                                                             NA
#> 6                                                                                                                             NA
#> 7  2008, FDZ_IUP, ISSP, ISSP_input2014, ISSP_pro, SCOPUSindexed, SSCIindexed, article, checked, indexproved, input2014, reviewed
#> 8                                                                                                                             NA
#> 9                                                                                                                             NA
#> 10                                                                                                                            NA
#>                                     study_group                   coreSseries
#> 1                                          <NA>                          <NA>
#> 2                                          <NA>                          <NA>
#> 3                                          <NA>                          <NA>
#> 4                                          <NA>                          <NA>
#> 5                                          <NA>                          <NA>
#> 6                                          <NA>                          <NA>
#> 7  International Social Survey Programme (ISSP)                          <NA>
#> 8                                          <NA>                          <NA>
#> 9                                          <NA> (Eurobarometer Spezial ; 490)
#> 10                                         <NA> (Eurobarometer Spezial ; 490)
#>                                                                                          intranet_link
#> 1                                                                                                 <NA>
#> 2                                                                                                 <NA>
#> 3                                                                                                 <NA>
#> 4                                                                                                 <NA>
#> 5                                                                                                 <NA>
#> 6                                                                                                 <NA>
#> 7                                                                                                 <NA>
#> 8                                                                                                 <NA>
#> 9  http://intranet.gesis.intra/Bibliothek/Volltexte/2019/Eurobarometer_2019_Spec490_Climate-change.pdf
#> 10 http://intranet.gesis.intra/Bibliothek/Volltexte/2019/Eurobarometer_2019_Spec490_Climate-change.pdf
#>    study_reference_numbers   study_references
#> 1                     <NA>               <NA>
#> 2                     <NA>               <NA>
#> 3                     <NA>               <NA>
#> 4                     <NA>               <NA>
#> 5                     <NA>               <NA>
#> 6                     <NA>               <NA>
#> 7                     <NA>               <NA>
#> 8                     <NA>               <NA>
#> 9                     7572 Eurobarometer 91.3
#> 10                    7572 Eurobarometer 91.3
gesis_search("climate change", tidy = TRUE)
#> # A tibble: 10 × 50
#>    id            title type  date  date_recency abstract portal_url topic person
#>    <chr>         <chr> <chr> <chr> <chr>        <chr>    <chr>      <lis> <list>
#>  1 csa-pais-201… Migr… publ… 2011  2011         "Provid… http://so… <chr> <chr> 
#>  2 gesis-ssoar-… Japa… publ… 2021  2021         "This a… https://w… <chr> <chr> 
#>  3 gesis-bib-14… The … gesi… NA    NA            NA      NA         <chr> <chr> 
#>  4 fes-bib-4691… Huma… publ… 2009  2009          NA      http://so… <chr> <chr> 
#>  5 gesis-bib-14… Futu… gesi… NA    NA            NA      NA         <chr> <chr> 
#>  6 gesis-ssoar-… Atti… publ… 2020  2020         "While … https://w… <chr> <chr> 
#>  7 bibsonomy-go… Clim… publ… 2008  2008         ""       NA         <chr> <chr> 
#>  8 gesis-bib-13… Intr… gesi… NA    NA            NA      NA         <chr> <chr> 
#>  9 gesis-bib-15… Clim… gesi… 2019  2019          NA      NA         <chr> <chr> 
#> 10 gesis-bib-15… Clim… publ… 2019  2019          NA      NA         <chr> <chr> 
#> # ℹ 41 more variables: person_sort <chr>, source <chr>, subtype <chr>,
#> #   publisher <chr>, database <chr>, document_type <chr>, coreAuthor <list>,
#> #   corePublisher <chr>, coreIsbn <chr>, coredoctypelitadd <chr>,
#> #   coreLanguage <chr>, data_source <chr>, link_count <chr>, gesis_own <int>,
#> #   fulltext <chr>, index_source <chr>, literature_collection <chr>,
#> #   coreSjahr <chr>, coreJournalTitle <chr>, coreZsband <chr>,
#> #   coreZsnummer <chr>, doi <chr>, coreIssn <chr>, metadata_quality <chr>, …

# Using the fields argument you can control how different text fields
# are weighted in matching the query string. Here we search for
# "climate change" only in full texts, not in other fields
gesis_search("climate change", fields = field_weights(full_text = 10))
#> A list of <gesis_hits> with 10 hits
#> 
#> <gesis_hit>
#> Type: publication
#> ID: gesis-ssoar-46462
#> Title: The Paris Agreement 2015: turning point for the international climate
#> regime
#> Date: 2016
#> Person:
#> • Dröge, Susanne
#> 
#> <gesis_hit>
#> Type: publication
#> ID: gesis-ssoar-22412
#> Title: Does tomorrow ever come? Disaster narrative and public perceptions of
#> climate change
#> Date: 2006
#> Persons:
#> • Lowe, Thomas
#> • Brown, Katrina
#> • Dessai, Suraje
#> • de França Doria, Miguel
#> • Haynes, Kat
#> • ... and 1 more
#> 
#> <gesis_hit>
#> Type: publication
#> ID: gesis-ssoar-19413
#> Title: Framing adaptation to climate change - a challenge for building
#> institutions
#> Date: 2008
#> Person:
#> • Horstmann, Britta
#> # ℹ 7 more hits
#> # ℹ Use `print(n = ...)` to see more hits

# Search strings are separated by an AND operator by default
# this default can be overwritten by setting default_operator = "OR"
# or by manually separating the query string
weight <- field_weights(title = 100)
and <- gesis_search("climate change", sort = "title_ascending", fields = weight)
or1 <- gesis_search("climate OR change", sort = "title_ascending", fields = weight)
or2 <- gesis_search("climate change", sort = "title_ascending", default_operator = "OR", fields = weight)
attr(and, "query") <- NULL
attr(or1, "query") <- NULL
attr(or2, "query") <- NULL
identical(and, or1)
#> [1] FALSE
identical(or1, or2)
#> [1] TRUE

# Similarly, you can construct complex query terms using operators
gesis_search("(climate or environmental) AND (change OR crisis OR action)")
#> A list of <gesis_hits> with 10 hits
#> 
#> <gesis_hit>
#> Type: publication
#> ID: gesis-ssoar-95723
#> Title: Environmental awareness: The case of climate change
#> Date: 2018
#> Persons:
#> • Weber, Shlomo
#> • Wiesmeth, Hans
#> 
#> <gesis_hit>
#> Type: publication
#> ID: bibsonomy-hoffmann2022climate
#> Title: Climate change experiences raise environmental concerns and promote
#> Green voting
#> Date: 2022
#> Persons:
#> • Hoffmann, Roman
#> • Muttarak, Raya
#> • Peisker, Jonas
#> • Stanig, Piero
#> 
#> <gesis_hit>
#> Type: publication
#> ID: gesis-ssoar-78644
#> Title: Urban Commons and Collective Action to Address Climate Change
#> Date: 2022
#> Persons:
#> • Colding, Johan
#> • Barthel, Stephan
#> • Ljung, Robert
#> • Eriksson, Felix
#> • Sjöberg, Stefan
#> # ℹ 7 more hits
#> # ℹ Use `print(n = ...)` to see more hits

# Another way to narrow down the search is to set filters. GESIS Search
# offers a number of filters, including topic, publication year or
# methodology. Sensible values can be checked on https://search.gesis.org/.
gesis_search(topics = "climate change", type = "publication") # search for records on climate change
#> A list of <gesis_hits> with 0 hits