GESIS Search

Query the GESIS search engine to retrieve metadata on research data, variables, tools, publications, libraries, and websites.

Usage

gesis_search(
  query = NULL,
  type = NULL,
  default_operator = "AND",
  fields = field_weights(),
  sort = "relevance",
  publication_year = NULL,
  countries = NULL,
  topics = NULL,
  authors = NULL,
  sources = NULL,
  study_title = NULL,
  study_group = NULL,
  collection_year = NULL,
  thematic_collections = NULL,
  method_data_collection = NULL,
  method_sampling = NULL,
  method_time = NULL,
  method_unit = NULL,
  data_type = NULL,
  interview_language = NULL,
  pages = 1,
  tidy = FALSE
)

# S3 method for class 'gesis_records'
as.data.frame(x, ...)

Arguments

query

The search term to query. Terms can be seperated by operators OR / AND. If no operator is specified, defaults to the operator specified in default_operator. If NULL, searches for all records. Queryless searches can still be filtered using the parameters below.

type

Type of search results to retrieve. Can be one of the following:

research_data: Survey data records

variables: Variables related to survey data records

instruments_tools: Measurement instruments like items, scales, questionnaires, tests, and indexes offered by ZIS.

publication: Publications like journal articles or book chapters

gesis_bib: Records from the GESIS library

gesis_web: GESIS websites

gesis_person: Institutional websites of GESIS employees

If NULL, includes all result types. Defaults to NULL.

default_operator

Character specifying the default operator to use as seperator of search terms. Can be "AND" or "OR". Defaults to "AND". Ignored if query is NULL.

fields

Character vector of field weights as returned by field_weights. Used to specify how much weight is given to query matches in certain metadata fields. The higher the weight, the more a field influences the match probability. Defaults to the default values specified by field_weights. Ignored if query is NULL.

sort

Attribute to sort by. Must be one of the following:

relevance: Sorts by relevance (whatever that means)

newest_first: Sorts by publication year, descending

oldest_first: Sorts by publication year, ascending

title_ascending: Sorts lexicographically by title, ascending

title_descending: Sorts lexicographically by title, descending

variable_ascending: Sorts lexicographically by variable name, ascending

variable_descending: Sorts lexicographically by variable name, descending

Defaults to "relevance" to mirror the behavior of the web interface.

publication_year

Numeric or character vector of length 2 specifying the time range in which a record was published. The first value is the start year and the second value is the end year. If NULL, returns records from all years. Defaults to NULL.

countries

Character vector specifying the country in which the record was published. If NULL, returns records from all countries. Defaults to NULL.

topics

Character vector specifying the topics to filter by. Example values include "Political behavior and attitudes" or "Migration". See https://search.gesis.org/ for a full list of possible topics.

authors

Character vector specifying the authors to filter by.

sources

Character vector specifying the source of the record.

study_title

Character vector specifying the study title. Only sensible for research data or variables.

study_group

Character vector specifying the study group. Only sensible for research data or variables.

collection_year

Numeric or character vector of length 2 specifying the time range in which a dataset was collected. The first value is the start year and the second value is the end year. If NULL, returns records from all years. Defaults to NULL. Only sensible for research data or variables.

thematic_collections

Character vector specifying the thematic collection a dataset or variable belongs to. Example values include "COVID-19" or "Environment and climate". See https://search.gesis.org/ for a full list of possible topics.

method_data_collection, method_sampling, method_time, method_unit

Arguments to filter by the methodology used in a study. It is possible to filter by the mode of data collection (method_data_collection), sampling procedure (method_sampling), temporal research design (method_unit) and analysis unit (method_unit). Only sensible for research data and variables.

data_type

Character vector specifying the data type of a dataset or variable. Example values include "numeric", "geospatial", and "text". Only sensible for research data and variables.

interview_language

Character vector specifying the interview language of a variable. Usually only "german" and "english" are available. Only sensible for variables.

pages

Numeric vector specifying the page numbers of the search results. By default, only the first page is returned. If NULL, iterates through all pages.

tidy

If TRUE, returns the records in a rectangular shape that can be wrangled more easily. In this form, some more complex metadata fields that cannot be converted to a rectangular shapre are discarded. If FALSE, returns the records as a list-like gesis_records object. This object is more comprehensive, but harder to work with. Defaults to FALSE.

x

Object of class gesis_records

...

Arguments passed on to the respective S3 methods.

Value

If tidy = FALSE, returns an object of class gesis_records consisting of multiple objects of class gesis_record. Each gesis_record contains attributes index, id, and score specifying the index, the record ID, and the match score.

If tidy = TRUE, returns a dataframe or tibble where each row is a record and each column is a metadata field.

In any case, the query used to retrieve the data is stored in the attribute query.

Examples

# Make a queryless search across the entire database and return
# the first 10 results
gesis_search()
#> A list of <gesis_records> with 10 records
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-44173
#> Title: Mujeres, historia y lucha
#> Date: 2010
#> Person:
#> • Zuluaga Muñoz, Deysy Alexandra
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-4419
#> Title: Rußlands Gazprom: T. 2, Gazprom als Akteur auf internationaler Ebene
#> Date: 1999
#> Person:
#> • Heinrich, Andreas
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-4425
#> Title: Probleme der Integration der baltischen Staaten in westliche Strukturen
#> Date: 1999
#> Person:
#> • Melnikas, Borisas
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-44252
#> Title: Ukraine 1994: Konturen eines neuen nationalen Projekts
#> Date: 1995
#> Person:
#> • Wittkowsky, Andreas
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-44300
#> Title: Probleme demokratischer Wiederbelebung in Peru: eine Zwischenbilanz der
#> Regierung Toledo
#> Date: 2002
#> Person:
#> • Steinhauf, Andreas
#> # ℹ 5 more records
#> # ℹ Use `print(n = ...)` to see more records

# Queries can be narrowed down by providing a query string
record <- gesis_search("climate change")

# gesis_search() searches for all types of records by default. If you only
# want to search for, say, datasets, you can set the `type` argument.
gesis_search(type = "research_data")
#> A list of <gesis_records> with 10 records
#> <gesis_record>
#> Type: research_data
#> ID: ZA0315
#> Title: IOS-Akademie
#> Date: 1968
#> Persons:
#> • Sommer, Klaus
#> • Stegemann, Hagen
#> 
#> <gesis_record>
#> Type: research_data
#> ID: ZA0340
#> Title: Olivetti-Vertreter (Fluktuation 1961)
#> Date: 1961
#> Person:
#> • DIVO, Frankfurt
#> 
#> <gesis_record>
#> Type: research_data
#> ID: ZA0343
#> Title: Firmenbild "Waschmittelindustrie"
#> Date: 1960
#> Person:
#> • DIVO, Frankfurt
#> 
#> <gesis_record>
#> Type: research_data
#> ID: ZA0344
#> Title: Beschäftigte in der Bauwirtschaft
#> Date: 1960
#> Person:
#> • DIVO, Frankfurt
#> 
#> <gesis_record>
#> Type: research_data
#> ID: ZA0360
#> Title: Probleme mittelständischer Unternehmer
#> Date: 1964
#> Persons:
#> • Daheim, Hans-Jürgen
#> • Kaupen, Wolfgang
#> # ℹ 5 more records
#> # ℹ Use `print(n = ...)` to see more records

# By default, only 10 results (1 page) are returned. You can specify
# which pages should queried or if all should be returned
gesis_search("climate change", pages = 2:3) # pages 2 and 3
#> A list of <gesis_records> with 20 records
#> <gesis_record>
#> Type: gesis_bib
#> ID: gesis-bib-126747
#> Title: Climate change and cultural diversity
#> Person:
#> • Wisner, Ben
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-79316
#> Title: Germany's Federal Climate Change Act
#> Date: 2021
#> Persons:
#> • Flachsland, Christian
#> • Levi, Sebastian
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-45173
#> Title: Can climate change negotiations succeed?
#> Date: 2013
#> Persons:
#> • Hovi, Jon
#> • Skodvin, Tora
#> • Aakre, Stine
#> 
#> <gesis_record>
#> Type: publication
#> ID: bibsonomy-flachsland2021germanys
#> Title: Germany’s Federal Climate Change Act
#> Date: 2021
#> Persons:
#> • Flachsland, Christian
#> • Levi, Sebastian
#> 
#> <gesis_record>
#> Type: publication
#> ID: bibsonomy-evansa2013climate
#> Title: Climate of Doubt: What Australians Think About Climate Change
#> Date: 2013
#> Persons:
#> • Evans, A.
#> • Reimondos, A.
#> # ℹ 15 more records
#> # ℹ Use `print(n = ...)` to see more records
if (FALSE) gesis_search("climate change", pages = NULL) # all pages

# By default, results are wrapped in a complex list structure. This is
# good for representing detailed information, but not for analyzing data.
# Results can be tidied up to a dataframe using either `as_tibble()`
# or by setting the `tidy` argument to TRUE
tibble::as_tibble(record)
#> # A tibble: 10 × 50
#>    id            title type  date  date_recency abstract portal_url topic person
#>    <chr>         <chr> <chr> <chr> <chr>        <chr>    <chr>      <lis> <list>
#>  1 csa-pais-201… Migr… publ… 2011  2011         "Provid… http://so… <chr> <chr> 
#>  2 gesis-bib-14… The … gesi… NA    NA            NA      NA         <chr> <chr> 
#>  3 gesis-ssoar-… Japa… publ… 2021  2021         "This a… https://w… <chr> <chr> 
#>  4 fes-bib-4691… Huma… publ… 2009  2009          NA      http://so… <chr> <chr> 
#>  5 gesis-bib-14… Futu… gesi… NA    NA            NA      NA         <chr> <chr> 
#>  6 gesis-bib-13… Intr… gesi… NA    NA            NA      NA         <chr> <chr> 
#>  7 gesis-ssoar-… Atti… publ… 2020  2020         "While … https://w… <chr> <chr> 
#>  8 gesis-bib-15… Clim… gesi… 2019  2019          NA      NA         <chr> <chr> 
#>  9 gesis-bib-15… Clim… publ… 2019  2019          NA      NA         <chr> <chr> 
#> 10 bibsonomy-go… Clim… publ… 2008  2008         ""       NA         <chr> <chr> 
#> # ℹ 41 more variables: person_sort <chr>, source <chr>, subtype <chr>,
#> #   publisher <chr>, database <chr>, document_type <chr>, coreAuthor <list>,
#> #   corePublisher <chr>, coreIsbn <chr>, coredoctypelitadd <chr>,
#> #   coreLanguage <chr>, data_source <chr>, link_count <chr>, gesis_own <int>,
#> #   fulltext <chr>, index_source <chr>, literature_collection <chr>,
#> #   coreJournalTitle <chr>, library_location <chr>, coreSjahr <chr>,
#> #   coreZsband <chr>, coreZsnummer <chr>, doi <chr>, coreIssn <chr>, …
gesis_search("climate change", tidy = TRUE)
#> # A tibble: 10 × 50
#>    id            title type  date  date_recency abstract portal_url topic person
#>    <chr>         <chr> <chr> <chr> <chr>        <chr>    <chr>      <lis> <list>
#>  1 csa-pais-201… Migr… publ… 2011  2011         "Provid… http://so… <chr> <chr> 
#>  2 gesis-bib-14… The … gesi… NA    NA            NA      NA         <chr> <chr> 
#>  3 gesis-ssoar-… Japa… publ… 2021  2021         "This a… https://w… <chr> <chr> 
#>  4 fes-bib-4691… Huma… publ… 2009  2009          NA      http://so… <chr> <chr> 
#>  5 gesis-bib-14… Futu… gesi… NA    NA            NA      NA         <chr> <chr> 
#>  6 gesis-bib-13… Intr… gesi… NA    NA            NA      NA         <chr> <chr> 
#>  7 gesis-ssoar-… Atti… publ… 2020  2020         "While … https://w… <chr> <chr> 
#>  8 gesis-bib-15… Clim… gesi… 2019  2019          NA      NA         <chr> <chr> 
#>  9 gesis-bib-15… Clim… publ… 2019  2019          NA      NA         <chr> <chr> 
#> 10 bibsonomy-go… Clim… publ… 2008  2008         ""       NA         <chr> <chr> 
#> # ℹ 41 more variables: person_sort <chr>, source <chr>, subtype <chr>,
#> #   publisher <chr>, database <chr>, document_type <chr>, coreAuthor <list>,
#> #   corePublisher <chr>, coreIsbn <chr>, coredoctypelitadd <chr>,
#> #   coreLanguage <chr>, data_source <chr>, link_count <chr>, gesis_own <int>,
#> #   fulltext <chr>, index_source <chr>, literature_collection <chr>,
#> #   coreJournalTitle <chr>, library_location <chr>, coreSjahr <chr>,
#> #   coreZsband <chr>, coreZsnummer <chr>, doi <chr>, coreIssn <chr>, …

# Using the fields argument you can control how different text fields
# are weighted in matching the query string.
# Search for "climate change" only in texts, not in other fields
gesis_search("climate change", fields = field_weights(
  full_text = 10,
  abstract_en = 5,
  abstract = 3
))
#> A list of <gesis_records> with 10 records
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-46462
#> Title: The Paris Agreement 2015: turning point for the international climate
#> regime
#> Date: 2016
#> Person:
#> • Dröge, Susanne
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-22412
#> Title: Does tomorrow ever come? Disaster narrative and public perceptions of
#> climate change
#> Date: 2006
#> Persons:
#> • Lowe, Thomas
#> • Brown, Katrina
#> • Dessai, Suraje
#> • de França Doria, Miguel
#> • Haynes, Kat
#> • ... and 1 more
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-61995
#> Title: The Baker's Dozen: Key Nations Can and Should Act Together to Prevent
#> Further Dangerous Climate Change
#> Date: 2009
#> Persons:
#> • Oleson, Kirsten
#> • Hartzell, Lauren
#> • Mastrandea, Michael D.
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-61737
#> Title: Scenario as a tool for critical thinking: climate change awareness and
#> denial as a case study
#> Date: 2018
#> Persons:
#> • Gramigna, Remo
#> • Marling, Raili
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-11793
#> Title: Local climate change policy in the United Kingdom and Germany
#> Date: 2004
#> Persons:
#> • Bulkeley, Harriet
#> • Kern, Kristine
#> # ℹ 5 more records
#> # ℹ Use `print(n = ...)` to see more records

# Search for articles in a specified journal
gesis_search("urban planning", fields = field_weights(coreJournalTitle = 10))
#> A list of <gesis_records> with 10 records
#> <gesis_record>
#> Type: publication
#> ID: bibsonomy-winters2024expanding
#> Title: Expanding the Social Rental Housing Stock in Flanders: Money Isn’t the
#> Problem
#> Date: 2024
#> Persons:
#> • Winters, Sien
#> • Dockx, Emmanuel
#> • den Broeck, Katleen Van
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-54401
#> Title: Re-Thinking Housing: From Physical Manifestation of Colonial Planning
#> Policy to Community-Focused Networks
#> Date: 2016
#> Person:
#> • McCartney, Shelagh
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-58364
#> Title: Evolving the Evolving: Territory, Place and Rewilding in the California
#> Delta
#> Date: 2017
#> Persons:
#> • Kraus-Polk, Alejo
#> • Milligan, Brett
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-58403
#> Title: Visualizing Conflict: Possibilities for Urban Research
#> Date: 2017
#> Persons:
#> • Saldarriaga, Juan Francisco
#> • Kurgan, Laura
#> • Brawley, Dare
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-58406
#> Title: Crowdsourced Quantification and Visualization of Urban Mobility Space
#> Inequality
#> Date: 2018
#> Person:
#> • Szell, Michael
#> # ℹ 5 more records
#> # ℹ Use `print(n = ...)` to see more records

# Search strings are separated by an AND operator by default
# this default can be overwritten by setting default_operator = "OR"
# or by manually separating the query string
weight <- field_weights(title = 100)
and <- gesis_search(
  "climate change",
  sort = "title_ascending",
  fields = weight
)
or1 <- gesis_search(
  "climate OR change",
  sort = "title_ascending",
  fields = weight
)
or2 <- gesis_search(
  "climate change",
  sort = "title_ascending",
  default_operator = "OR",
  fields = weight
)
attr(and, "query") <- NULL
attr(or1, "query") <- NULL
attr(or2, "query") <- NULL
identical(and, or1)
#> [1] FALSE
identical(or1, or2)
#> [1] TRUE

# Similarly, you can construct complex query terms using operators
gesis_search("(climate or environmental) AND (change OR crisis OR action)")
#> A list of <gesis_records> with 10 records
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-95723
#> Title: Environmental awareness: The case of climate change
#> Date: 2018
#> Persons:
#> • Weber, Shlomo
#> • Wiesmeth, Hans
#> 
#> <gesis_record>
#> Type: publication
#> ID: bibsonomy-uysal2024system
#> Title: “System Change, not Climate Change”: Effective Environmental Policies
#> and State Repression Moderate the Relationship Between Psychological Predictors
#> and Environmental Collective Action
#> Date: 2024
#> Persons:
#> • Uysal, Mete Sefa
#> • Vestergren, Sara
#> • Varela, Micaela
#> • Lindner, Clemens
#> 
#> <gesis_record>
#> Type: publication
#> ID: bibsonomy-hoffmann2022climate
#> Title: Climate change experiences raise environmental concerns and promote
#> Green voting
#> Date: 2022
#> Persons:
#> • Hoffmann, Roman
#> • Muttarak, Raya
#> • Peisker, Jonas
#> • Stanig, Piero
#> 
#> <gesis_record>
#> Type: publication
#> ID: gesis-ssoar-78644
#> Title: Urban Commons and Collective Action to Address Climate Change
#> Date: 2022
#> Persons:
#> • Colding, Johan
#> • Barthel, Stephan
#> • Ljung, Robert
#> • Eriksson, Felix
#> • Sjöberg, Stefan
#> 
#> <gesis_record>
#> Type: publication
#> ID: bibsonomy-tranterbruce2011political
#> Title: Political divisions over climate change and environmental issues in
#> Australia
#> Date: 2011
#> Person:
#> • Tranter, Bruce
#> # ℹ 5 more records
#> # ℹ Use `print(n = ...)` to see more records

# Another way to narrow down the search is to set filters. GESIS Search
# offers a number of filters, including topic, publication year or
# methodology. Sensible values can be checked on https://search.gesis.org/.
gesis_search(topics = "climate change", type = "publication")
#> A list of <gesis_records> with 0 records